Task 04 (phase 2): extract :core:model from :core
Move FieldText into a new :core:model module (depends on :core:common for
MviError). Rewrite FieldText imports in :feature:auth and :core.

Fix a cross-module smart-cast in TextFieldWithError: FieldText.error is now a
public API property in another module, so capture it in a local val before the
null check.

Builds green: assembleDebug + detektDebug + ktlintCheck.
