+++ core/build.gradle.kts
@@ -14,6 +14,7 @@ android {
 
 dependencies {
     implementation(project(":core:common"))
+    implementation(project(":core:model"))
 
     implementation(platform(libs.androidx.compose.bom))
