+++ feature/auth/src/main/java/com/timhome/auth/ui/signin/LoginViewState.kt
@@ -1,7 +1,7 @@
 package com.timhome.auth.ui.signin
 
 import android.os.Parcelable
-import com.timhome.core.data.models.FieldText
+import com.timhome.core.model.FieldText
 import com.timhome.core.common.mvi.MviError
 import com.timhome.core.common.mvi.MviViewState
 import kotlinx.parcelize.Parcelize
